Comments, layout editor now uses reasonable DnD interface.
Grabbing a widget in the layout editor will now cause the DnD
image (the pixbuf) to "lock" to the cursor wherever it was grabbed
rather than locking to the upper-left corner. There is still a
little weirdness with he GTK DnD behavior having a "delay" in grabbing
windows. Try grabbing a widget with a quickly-moving pointer
for a better idea of what this means.