Fix Wah!Cade layout editor to properly read new YAML format layout files.
The layout editor now properly reads and writes the new YAML-based
layout files, which replace the nearly unreadable super-array format.
The new format has extension .layy
If a .layy file is not detected when attempting to load, the legacy
loader is invoked and the program attempts to load a .lay file if present.