Add new Popular Games list via list-switching functionality.
ITP-76
Pressing "Next List" (default '3') will eventually lead to the new
Popular Games list in the Main window. It is populated via a new list
type (xml_remote), which allows for arbitrary calls to a server,
as specified in the associated .ini file, under the assumption that
XML is returned. The XML is parsed, and fed into the new underlying
list construct to populate the new list display.