Fix games not highlighting correctly on new lists.
ITP-166
The file being used to store the generated list of games was being
repeatedly opened in utf-8-sig format, which caused an extraneous
BOM to be added to the file on each append operation. This resulted
in the rom name being stored incorrectly, and thus failing the
lookup against the main games list.
The file is now tested for presence before opening it; if it already
exists it is opened in the default encoding mode which does not
add a BOM.