Fix program hang when switching to empty game list.
ITP-231
The program was searching a 2000+ file directory once for each
of about 6 filetypes once for each of a couple filename patterns.
This produced a noticeable delay in switching to an empty game
list due to no file being found and shortcircuiting the search.
By testing for an empty rom name in wc_common.get_artwork_image,
this useless search loop is avoided.

Additionally, a couple if statements were made more pythonic.